:sparkles: Fixed ``interpretDynamicCommandsArray()``, divised ``getCommandValue()`` into 2 functions, added function call return value re-interpreation in it
I also managed to get a decent test through with it. A function call with 4 arguments - First is a string, second is a uint256, third is another FunctionCall to a function which has a dynamic array of YC commands as it's only argument, where the first one in it is a string (dynamic variable command), which returns the string concatinated with another string (So the underlying type is a string i.e dynamic variable). And it works!